Feature: Support one to many

+++ DataEditorPortal.Web/Services/IQueryBuilder/IQueryBuilder.cs
@@ -34,9 +34,11 @@ namespace DataEditorPortal.Web.Services
         string GenerateSqlTextForUpdate(DataSourceConfig config);
         string GenerateSqlTextForDelete(DataSourceConfig config);
         string GenerateSqlTextForExist(DataSourceConfig config);
-        string GenerateSqlTextForLinkData(TableMeta linkTable, TableMeta t1, TableMeta t2);
-        string GenerateSqlTextForDeleteLinkData(DataSourceConfig config, TableMeta input);
+        string GenerateSqlTextForLinkData(TableMeta t1, TableMeta t2, string queryAllRelations = null, bool inputIsPrimary = true);
+        string GenerateSqlTextForDeleteLinkData(RelationDataSourceConfig config, TableMeta input, TableMeta output);
         string GenerateSqlTextForQueryForeignKeyValue(TableMeta input);
+        string GenerateSqlTextForRemoveRelation(DataSourceConfig config, bool isOneToMany);
+        string GenerateSqlTextForAddRelation(DataSourceConfig config, bool isOneToMany);
 
         string GenerateSqlTextForColumnFilterOption(DataSourceConfig config);
 
@@ -415,42 +417,77 @@ namespace DataEditorPortal.Web.Services
             return $@"SELECT {EscapeColumnName(config.IdColumn)} FROM ({queryText}) A WHERE {EscapeColumnName(config.IdColumn)} IN {ParameterPrefix}{ParameterName(config.IdColumn)}";
         }
 
-        /// <summary>
-        /// get the link data by table1 id
-        /// </summary>
-        /// <param name="dsLink"></param>
-        /// <param name="input"></param>
-        /// <param name="output"></param>
-        /// <returns></returns>
-        public virtual string GenerateSqlTextForLinkData(TableMeta linkTable, TableMeta input, TableMeta output)
+
+        #region query for link table
+
+        public virtual string GenerateSqlTextForLinkData(TableMeta input, TableMeta output, string queryAllRelations = null, bool inputIsPrimary = true)
         {
-            var linkQuery = $@"
+            if (!string.IsNullOrEmpty(queryAllRelations))
+            {
+                // many to many
+                return $@"
                     SELECT 
-                        link.{EscapeColumnName(linkTable.IdColumn)} AS ""LINK_{linkTable.IdColumn}"", 
                         t1.{EscapeColumnName(input.IdColumn)} AS ""T1_{input.IdColumn}"", 
                         t2.{EscapeColumnName(output.IdColumn)} AS ""T2_{output.IdColumn}"",
                         link.{EscapeColumnName(input.ForeignKey)} AS ""F1_{input.ForeignKey}"", 
                         link.{EscapeColumnName(output.ForeignKey)} AS ""F2_{output.ForeignKey}""
                     FROM ({input.Query_AllData}) t1
-                    INNER JOIN ({linkTable.Query_AllData}) link ON t1.{EscapeColumnName(input.ReferenceKey)} = link.{EscapeColumnName(input.ForeignKey)} 
+                    INNER JOIN ({queryAllRelations}) link ON t1.{EscapeColumnName(input.ReferenceKey)} = link.{EscapeColumnName(input.ForeignKey)} 
                     INNER JOIN ({output.Query_AllData}) t2 ON t2.{EscapeColumnName(output.ReferenceKey)} = link.{EscapeColumnName(output.ForeignKey)} 
                     WHERE t1.{input.IdColumn} = {ParameterPrefix}{ParameterName(input.IdColumn)}
                 ";
-
-            return linkQuery;
+            }
+            else
+            {
+                if (inputIsPrimary)
+                {
+                    return $@"
+                        SELECT 
+                            t1.{EscapeColumnName(input.IdColumn)} AS ""T1_{input.IdColumn}"", 
+                            t2.{EscapeColumnName(output.IdColumn)} AS ""T2_{output.IdColumn}""
+                        FROM ({input.Query_AllData}) t1
+                        INNER JOIN ({output.Query_AllData}) t2 ON t1.{EscapeColumnName(output.ReferenceKey)} = t2.{EscapeColumnName(output.ForeignKey)} 
+                        WHERE t1.{input.IdColumn} = {ParameterPrefix}{ParameterName(input.IdColumn)}
+                    ";
+                }
+                else
+                {
+                    return $@"
+                        SELECT 
+                            t1.{EscapeColumnName(input.IdColumn)} AS ""T1_{input.IdColumn}"", 
+                            t2.{EscapeColumnName(output.IdColumn)} AS ""T2_{output.IdColumn}""
+                        FROM ({input.Query_AllData}) t1
+                        INNER JOIN ({output.Query_AllData}) t2 ON t2.{EscapeColumnName(output.ReferenceKey)} = t1.{EscapeColumnName(output.ForeignKey)} 
+                        WHERE t1.{input.IdColumn} = {ParameterPrefix}{ParameterName(input.IdColumn)}
+                    ";
+                }
+            }
         }
 
-        public virtual string GenerateSqlTextForDeleteLinkData(DataSourceConfig config, TableMeta input)
+        public virtual string GenerateSqlTextForDeleteLinkData(RelationDataSourceConfig config, TableMeta input, TableMeta output)
         {
-            return $@"
-                DELETE FROM {config.TableSchema}.{config.TableName}
-                WHERE {EscapeColumnName(input.ForeignKey)} IN (
-                    SELECT link.{EscapeColumnName(input.ForeignKey)} 
-                    FROM {config.TableSchema}.{config.TableName} link
-                    INNER JOIN ({input.Query_AllData}) t1 ON link.{EscapeColumnName(input.ForeignKey)} = t1.{EscapeColumnName(input.ReferenceKey)}
-                    WHERE t1.{input.IdColumn} IN {ParameterPrefix}{ParameterName(input.IdColumn)}
-                )
-            ";
+            if (config.IsOneToMany)
+            {
+                return $@"
+                    UPDATE {output.TableSchema}.{output.TableName} SET {output.ForeignKey} = NULL 
+                    WHERE {output.ForeignKey} IN (
+                        SELECT {input.ReferenceKey} FROM ({input.Query_AllData}) t1
+                        WHERE t1.{input.IdColumn} IN {ParameterPrefix}{ParameterName(input.IdColumn)}
+                    )
+                ";
+            }
+            else
+            {
+                return $@"
+                    DELETE FROM {config.TableSchema}.{config.TableName}
+                    WHERE {EscapeColumnName(input.ForeignKey)} IN (
+                        SELECT link.{EscapeColumnName(input.ForeignKey)} 
+                        FROM {config.TableSchema}.{config.TableName} link
+                        INNER JOIN ({input.Query_AllData}) t1 ON link.{EscapeColumnName(input.ForeignKey)} = t1.{EscapeColumnName(input.ReferenceKey)}
+                        WHERE t1.{input.IdColumn} IN {ParameterPrefix}{ParameterName(input.IdColumn)}
+                    )
+                ";
+            }
         }
 
         public virtual string GenerateSqlTextForQueryForeignKeyValue(TableMeta input)
@@ -461,6 +498,62 @@ namespace DataEditorPortal.Web.Services
 
             return $@"SELECT {string.Join(",", columns)} FROM ({input.Query_AllData}) A WHERE {EscapeColumnName(input.IdColumn)} IN {ParameterPrefix}{ParameterName(input.IdColumn)}";
         }
+
+        public virtual string GenerateSqlTextForAddRelation(DataSourceConfig config, bool isOneToMany)
+        {
+            if (isOneToMany)
+            {
+                //var sets = string.Join(", ", config.Columns.Select(x => $"{EscapeColumnName(x)}={ParameterPrefix}{ParameterName(x)}"));
+
+                //var source = string.IsNullOrEmpty(config.TableName) ? config.TableName : $"{config.TableSchema}.{config.TableName}";
+                //return $@"UPDATE {source} SET {sets} WHERE {EscapeColumnName(config.IdColumn)} = {ParameterPrefix}{ParameterName(config.IdColumn)}";
+                return GenerateSqlTextForUpdate(config);
+            }
+            else
+            {
+                //return GenerateSqlTextForInsert(config);
+                if (!string.IsNullOrEmpty(config.QueryText))
+                {
+                    return ReplaceQueryParamters(config.QueryText);
+                }
+                else
+                {
+                    if (config.Columns.Count <= 0) throw new Exception("Columns can not be empty during generating insert script.");
+
+                    var source = string.IsNullOrEmpty(config.TableName) ? config.TableName : $"{config.TableSchema}.{config.TableName}";
+
+                    var columns = string.Join(",", config.Columns.Select(x => EscapeColumnName(x)));
+
+                    var param = string.Join(",", config.Columns.Select(x => $"{ParameterPrefix}{ParameterName(x)}"));
+
+                    var queryText = $@"INSERT INTO {source} ({columns}) VALUES ({param})";
+
+                    return queryText;
+                }
+            }
+        }
+        public virtual string GenerateSqlTextForRemoveRelation(DataSourceConfig config, bool isOneToMany)
+        {
+            if (isOneToMany)
+            {
+                //var sets = string.Join(", ", config.Columns.Select(x => $"{EscapeColumnName(x)}=NULL"));
+
+                //var source = string.IsNullOrEmpty(config.TableName) ? config.TableName : $"{config.TableSchema}.{config.TableName}";
+                //return $@"UPDATE {source} SET {sets} WHERE {EscapeColumnName(config.IdColumn)} = {ParameterPrefix}{ParameterName(config.IdColumn)}";
+                return GenerateSqlTextForUpdate(config);
+            }
+            else
+            {
+                var where = config.Filters.Select(x => { return $@"{EscapeColumnName(x.field)} = {ParameterPrefix}{ParameterName(x.field)}"; });
+                if (where.Count() < 2) throw new DepException("Two criterias are required when remove Many to Many relation.");
+
+                var source = string.IsNullOrEmpty(config.TableName) ? config.TableName : $"{config.TableSchema}.{config.TableName}";
+                return $@"DELETE FROM {source} WHERE {string.Join(" AND ", where)}";
+            }
+        }
+
+        #endregion
+
         #endregion
 
         public (string, List<KeyValuePair<string, object>>) ProcessQueryWithParamters(string queryText, IDictionary<string, object> model)
